statiticalTest 测试

+++ operators/statistical-test/src/main/java/com/bdilab/dataflow/dto/TdisDescription.java
@@ -16,10 +16,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class TdisDescription {
   private double meanX1;
-  private Integer n1;
+  private Long n1;
   private double sampleVariance1;  //sampleVariance = std^2
 
   private double meanX2;
-  private Integer n2;
+  private Long n2;
   private double sampleVariance2;
 }
